requirements.txt
dlib==19.15.0
Flask==1.0.2
Flask-SocketIO==3.0.1
h5py==2.8.0
imutils==0.5.1
Keras==2.2.2
matplotlib==2.2.3
numpy==1.14.5
opencv-python==3.4.2.17
Pillow==5.2.0
tensorflow==1.10.0
